rename: weather 패키지 내 service 코드 폴더 구조화
