+++ src/main/java/com/example/demo/weather/service/fcst/MidFcstService.java
@@ -1,4 +1,4 @@
-package com.example.demo.weather.service;
+package com.example.demo.weather.service.fcst;
 
 import com.example.demo.weather.dto.mid_land.MidLandApiResponse;
 import com.example.demo.weather.dto.mid_ta.MidTaApiResponse;
