+++ src/main/java/com/example/demo/weather/service/fcst/SrtFcstApiService.java
@@ -1,4 +1,4 @@
-package com.example.demo.weather.service;
+package com.example.demo.weather.service.fcst;
 
 import com.example.demo.weather.dto.srt_fcst.FcstApiResponse;
 import com.example.demo.weather.dto.srt_ncst.NcstApiResponse;
