+++ src/main/java/com/example/demo/weather/service/ncp/NcpMapApiService.java
@@ -1,4 +1,4 @@
-package com.example.demo.weather.service;
+package com.example.demo.weather.service.ncp;
 
 import com.example.demo.weather.dto.geocoding.GeoNcpResponse;
